ISIS-3223: move around classes/packages
- separation of domain code from test code
