--- regressiontests/stable-core-wrapperfactory/logging-dn-enhance.properties
@@ -1,48 +1,0 @@
-#
-#  Licensed to the Apache Software Foundation (ASF) under one
-#  or more contributor license agreements.  See the NOTICE file
-#  distributed with this work for additional information
-#  regarding copyright ownership.  The ASF licenses this file
-#  to you under the Apache License, Version 2.0 (the
-#  "License"); you may not use this file except in compliance
-#  with the License.  You may obtain a copy of the License at
-#
-#        http://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing,
-#  software distributed under the License is distributed on an
-#  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-#  KIND, either express or implied.  See the License for the
-#  specific language governing permissions and limitations
-#  under the License.
-#
-
-log4j.appender.A1=org.apache.log4j.FileAppender
-log4j.appender.A1.File=datanucleus.log
-log4j.appender.A1.layout=org.apache.log4j.PatternLayout
-log4j.appender.A1.layout.ConversionPattern=%d{HH:mm:ss,SSS} (%t) %-5p [%c] - %m%n
-
-
-# overriding all those below... 
-log4j.category.DataNucleus=ERROR
-
-log4j.category.DataNucleus.Persistence=INFO, A1
-log4j.category.DataNucleus.Transaction=INFO, A1
-log4j.category.DataNucleus.Connection=INFO, A1
-log4j.category.DataNucleus.Query=INFO, A1
-log4j.category.DataNucleus.Cache=INFO, A1
-log4j.category.DataNucleus.MetaData=INFO, A1
-log4j.category.DataNucleus.Datastore=INFO, A1
-log4j.category.DataNucleus.Datastore.Schema=INFO, A1
-log4j.category.DataNucleus.Datastore.Persist=INFO, A1
-log4j.category.DataNucleus.Datastore.Retrieve=INFO, A1
-#Log of all 'native' statements sent to the datastore
-log4j.category.DataNucleus.Datastore.Native=INFO, A1 
-log4j.category.DataNucleus.General=INFO, A1
-#All messages relating to object lifecycle changes
-log4j.category.DataNucleus.Lifecycle=INFO, A1
-log4j.category.DataNucleus.ValueGeneration=INFO, A1
-log4j.category.DataNucleus.Enhancer=INFO, A1
-log4j.category.DataNucleus.SchemaTool=INFO, A1
-log4j.category.DataNucleus.JDO=INFO, A1
-
