+++ regressiontests/stable/src/main/java/org/apache/isis/testdomain/wrapperfactory/Counter_bumpUsingMixin.java
@@ -18,7 +18,7 @@
  *
  */
 
-package org.apache.isis.regressiontests.core.wrapperfactory.integtests.model;
+package org.apache.isis.testdomain.wrapperfactory;
 
 import lombok.RequiredArgsConstructor;
